Python/Tooling/WebScraping/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
openai
faiss-cpu
tiktoken
python-multipart
pydantic
pytesseract
Pillow
